QCB-5-User Moved two files and implemented getUserById feature
